+++ apps/runtime-demo/3005-runtime-host/@mf-types/remote1/WebpackPng.d.ts
@@ -1,0 +1,2 @@
+export * from './compiled-types/apps/runtime-demo/3006-runtime-remote/src/components/WebpackPng';
+export { default } from './compiled-types/apps/runtime-demo/3006-runtime-remote/src/components/WebpackPng';