+++ packages/dts-plugin/src/plugins/utils.ts
@@ -14,3 +14,7 @@ export const isTSProject = (tsConfigPath?: string, context = process.cwd()) => {
     return false;
   }
 };
+
+export function isDev(): boolean {
+  return process.env['NODE_ENV'] === 'development';
+}