+++ packages/dts-plugin/vite.config.ts
@@ -1,0 +1,20 @@
+/// <reference types="vitest" />
+import { defineConfig } from 'vite';
+
+import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
+
+export default defineConfig({
+  cacheDir: '../../node_modules/.vite/native-federation-tests',
+
+  plugins: [nxViteTsPaths()],
+
+  test: {
+    cache: {
+      dir: '../../node_modules/.vitest',
+    },
+    environment: 'node',
+    include: ['src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
+    reporters: ['default'],
+    testTimeout: 60000,
+  },
+});