+++ packages/enhanced/src/lib/container/ModuleFederationPlugin.ts
@@ -15,6 +15,7 @@ import ContainerPlugin from './ContainerPlugin';
 import ContainerReferencePlugin from './ContainerReferencePlugin';
 import schema from '../../schemas/container/ModuleFederationPlugin';
 import FederationRuntimePlugin from './runtime/FederationRuntimePlugin';
+import HoistContainerReferencesPlugin from './HoistContainerReferencesPlugin';
 
 const isValidExternalsType = require(
   normalizeWebpackPath(
@@ -52,6 +53,7 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
    */
   apply(compiler: Compiler): void {
     const { _options: options } = this;
+    // @ts-ignore
     new FederationRuntimePlugin(options).apply(compiler);
     const library = options.library || { type: 'var', name: options.name };
     const remoteType =
@@ -70,8 +72,6 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
     if (useContainerPlugin) {
       // @ts-ignore
       ContainerPlugin.patchChunkSplit(compiler, this._options.name);
-      ContainerPlugin.patchChunkSplit(compiler, 'federation-runtime');
-      ContainerPlugin.patchChunkSplit(compiler, 'mfp-runtime-plugins');
     }
 
     if (!disableManifest && useContainerPlugin) {
@@ -108,6 +108,7 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
           //@ts-ignore
           exposes: options.exposes,
           runtimePlugins: options.runtimePlugins,
+          //@ts-ignore
         }).apply(compiler);
       }
       if (