+++ packages/native-federation-typescript/src/interfaces/HostOptions.ts
@@ -1,15 +1,6 @@
-import { moduleFederationPlugin } from '@module-federation/sdk';
-
-export interface HostOptions extends moduleFederationPlugin.DtsHostOptions {
-  moduleFederationConfig: moduleFederationPlugin.ModuleFederationPluginOptions;
-  context?: string;
-  implementation?: string;
-}
-
-export interface RemoteInfo {
-  name: string;
-  url: string;
-  alias: string;
-  zipUrl?: string;
-  apiTypeUrl?: string;
+export interface HostOptions {
+  moduleFederationConfig: any;
+  typesFolder?: string;
+  deleteTypesFolder?: boolean;
+  maxRetries?: number;
 }