+++ packages/rspack/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @module-federation/rspack
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [f926b6c]
+  - @module-federation/dts-plugin@0.1.3
+  - @module-federation/manifest@0.1.3
+  - @module-federation/runtime-tools@0.1.3
+  - @module-federation/sdk@0.1.3
+  - @module-federation/managers@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes